ai/requirements.txt

PyMuPDF
sentence-transformers
pymongo
python-dotenv
requests
flask